- Fixing duplication
