+++ app/tests/views/test_gym_item_view.py
@@ -1,26 +1,12 @@
 """ Test the gym item view """
 from datetime import datetime
-from django.test import TestCase
-from django.contrib.auth.models import User
+from gym_item_common import GymViewCommonCase
 from django.core.urlresolvers import reverse_lazy
-from app.tests.common import create_gym_item
 from app.models.gym_item import GymItem
 
 
-class TestGymItemView(TestCase):
+class TestGymItemView(GymViewCommonCase):
     """ Gym Item view tests """
-
-    def setUp(self):
-        """ Set up User for testing """
-        super(TestGymItemView, self).setUp()
-        create_gym_item('test', 'Test Gym', 'Test Location', '1988-01-12')
-        create_gym_item(
-            'new_user', 'Test Gym 1', 'Test Location', '1988-01-12')
-        self.user = User.objects.get(username='test')
-        self.gym_item = GymItem.objects.get(gym__name='Test Gym')
-        self.other_gym_item = GymItem.objects.get(gym__name='Test Gym 1')
-        self.user.set_password('password')
-        self.user.save()
 
     def test_redirects_logged_out_user(self):
         """ Test that a logged out user is redirected to the Homepage """
