+++ app/tests/views/test_gym_list_view.py
@@ -1,21 +1,11 @@
 """ Test the gym list view """
-from django.test import TestCase
-from django.contrib.auth.models import User
+from gym_item_common import GymViewCommonCase
 from django.core.urlresolvers import reverse_lazy
-from app.tests.common import create_gym_item
 from app.models.gym_item import GymItem
 
 
-class TestGymListView(TestCase):
+class TestGymListView(GymViewCommonCase):
     """ Gym List view tests """
-
-    def setUp(self):
-        """ Set up User for testing """
-        super(TestGymListView, self).setUp()
-        create_gym_item('test', 'Test Gym', 'Test Location', '1988-01-12')
-        self.user = User.objects.get(username='test')
-        self.user.set_password('password')
-        self.user.save()
 
     def test_redirects_logged_out_user(self):
         """ Test that a logged out user is redirected to the Homepage """
